+++ vm/src/hint_processor/cairo_1_hint_processor/dict_manager.rs
@@ -3,6 +3,7 @@ use num_traits::One;
 use crate::stdlib::collections::HashMap;
 use crate::stdlib::prelude::*;
 
+use crate::types::relocatable::MaybeRelocatable;
 use crate::vm::errors::hint_errors::HintError;
 use crate::Felt252;
 use crate::{types::relocatable::Relocatable, vm::vm_core::VirtualMachine};
@@ -10,7 +11,7 @@ use crate::{types::relocatable::Relocatable, vm::vm_core::VirtualMachine};
 /// Stores the data of a specific dictionary.
 pub struct DictTrackerExecScope {
     /// The data of the dictionary.
-    data: HashMap<Felt252, Felt252>,
+    data: HashMap<Felt252, MaybeRelocatable>,
     /// The index of the dictionary in the dict_infos segment.
     #[allow(dead_code)]
     idx: usize,
@@ -80,13 +81,22 @@ impl DictManagerExecScope {
     }
 
     /// Inserts a value to the dict tracker corresponding to a given pointer to a dict segment.
-    pub fn insert_to_tracker(&mut self, dict_end: Relocatable, key: Felt252, value: Felt252) {
+    pub fn insert_to_tracker(
+        &mut self,
+        dict_end: Relocatable,
+        key: Felt252,
+        value: MaybeRelocatable,
+    ) {
         self.get_dict_tracker_mut(dict_end).data.insert(key, value);
     }
 
     /// Gets a value from the dict tracker corresponding to a given pointer to a dict segment.
     /// None if the key does not exist in the tracker data.
-    pub fn get_from_tracker(&self, dict_end: Relocatable, key: &Felt252) -> Option<Felt252> {
+    pub fn get_from_tracker(
+        &self,
+        dict_end: Relocatable,
+        key: &Felt252,
+    ) -> Option<MaybeRelocatable> {
         self.get_dict_tracker(dict_end).ok()?.data.get(key).cloned()
     }
 }